Refactor filtering
While looking into how we could extend the filtering with other options (like nearest leader), I thought that the current filtering logic was needlessly complex. Before this PR, when a filtering happened, the state of rows was overwritten with the matching rows, and the original state of rows was lost (called `allRows`). To accomodate for this, the prop `allRows` was passed to _alot_ of components which really didn't have anything to do with it, other than having it available when calling `dispatch` (which was also passed around). Also, the reducer-logic was kinda hard to wrap my head around, considering each case called a function which accepted 8 arguments, each either representing the current state or the value of the action called.

So to make this a little less complex, I've done the following:

1. Removed all the complex reducer-logic and replaced it with a single `useState` object: `filter`. This contains all the current filtering. When for example a search is made, we update `filter.searchTerm` with the corresponding value (with `setFilter`, of course). Note that I considered using a reducer to handle the filtering, but considering it's just one level deep right now, I don't think the extra code required makes it worth it.
2. Instead of having to keep track of the original state by passing `allRows` all over the place, we now do the filtering when the component renders, based on what's in the `filter` object. The value of `filteredRows` is then what's displayed in the table.
3. There was some hardcoded logic in namely `FilterHeader` and `CompetenceFilterInput` based on their type: `'COMPETENCE' | 'MOTIVATION'`. Based on these values, they did different things, like call different dispatches or have different placeholders. I've changed these components to instead have prop functions like `onSelect` and `onThresholdUpdate`, which are called from inside the component. The parentcomponent can then decide what to do when they're called. There was also some hardcoded logic in `SearchInput` which has been replaced in the same manner.