Remove flash of login screen for logged in users
A returning user will have a previously set access token in localStorage. If such a token is present we let the UserInfo remain in undefined state for up to 3 seconds to allow the user info to be fetched.

For users that were not logged in in previous session we transition from undefined to null immediately and serve the login page.